Document Endianness of Varints
VarInts can be encoded as either big or little-endian. With the latest expansion adding cards that have codes requiring more than a single byte, there were some encoding errors in some of the deck code libraries that were using little-endian encodings. There doesn't seem to be anywhere in the docs that specifies that the VarInts are big-endian so I'd like to add it to clear up any confusion in the future.